document-portal: Rename main sources
xdp-fuse.c -> document-portal-fuse.c
xdp-fuse.h -> document-portal-fuse.h
xdp-main.c -> document-portal.c

The "xdp" term is soo overloaded.
